refactor: reorganize Zustand stores into domain-based structure
- Moved project and user stores under entities/
- Moved modal-related store under features/manage-modal
- Improved overall state organization and consistency
